Add caching to JVM info factory
