+++ dd-java-agent/agent-ci-visibility/src/test/groovy/datadog/trace/civisibility/config/JvmInfoFactoryTest.groovy
@@ -15,7 +15,7 @@ class JvmInfoFactoryTest extends Specification {
     def currentJvmExecutable = getCurrentJvmExecutable()
 
     when:
-    def jvmInfo = JvmInfoFactory.doGetJvmInfo(currentJvmExecutable)
+    def jvmInfo = JvmInfoFactoryImpl.doGetJvmInfo(currentJvmExecutable)
 
     then:
     jvmInfo == JvmInfo.CURRENT_JVM
